+++ core/src/main/java/org/sonatype/goodies/dropwizard/app/ApplicationCustomizer.java
@@ -10,7 +10,7 @@
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the Apache License Version 2.0 for the specific language governing permissions and limitations there under.
  */
-package org.sonatype.goodies.dropwizard;
+package org.sonatype.goodies.dropwizard.app;
 
 import java.util.Collections;
 import java.util.List;
