+++ core/src/main/java/org/sonatype/goodies/dropwizard/internal/ComponentDiscovery.java
@@ -24,7 +24,7 @@ import javax.ws.rs.container.DynamicFeature;
 import javax.ws.rs.core.Feature;
 import javax.ws.rs.ext.ExceptionMapper;
 
-import org.sonatype.goodies.dropwizard.EnvironmentCustomizer;
+import org.sonatype.goodies.dropwizard.env.EnvironmentCustomizer;
 import org.sonatype.goodies.dropwizard.jaxrs.Component;
 import org.sonatype.goodies.dropwizard.jaxrs.Resource;
