+++ core/src/main/resources/META-INF/services/io.dropwizard.jackson.Discoverable
@@ -1,1 +1,1 @@
-org.sonatype.goodies.dropwizard.ConfigurationAttachment
+org.sonatype.goodies.dropwizard.config.ConfigurationAttachment
